Update FilamentReader.ino
Expand firmware capabilities to allow data logging to be started, stopped and reset via the serial console. This is achieved by sending B, S or R to the device.

Further, sending I <value> will adjust the polling rate when data is being returned.

Returned data is now in the format of:
<distance_since_last>,<total_distance>